add login with spotify option - includes access, refresh tokens, requests, logic for spotify tokens with scope, new env variables for backend -,  remove username from Account tab in navbar, delete playlists by playlistId instead of _id because sets up future change where we can submit playlist without an account - also needed for new spotify playlists -, new spotifyPlaylistTile where we can submit playlistid without saving to the database which is needed for these personal spotify playlists,  new localstorage methods,  properly clear jwt on logoutby sending get request to logout to backend when logging out ,  handle edge case where artistGenres is empty - genre defaults to 'Unknown'

+++ client/src/components/PlaylistContainer/PlaylistContainer.jsx
@@ -26,6 +26,7 @@ const PlaylistContainer = ({ playlist, refreshPlaylist }) => {
   const [deleteLoader, setDeleteLoader] = useState(false)
   const { 
     _id: playlistObjectId,
+    playlistId,
     playlistName,
     playlistOwner,
     playlistImage,
@@ -40,20 +41,23 @@ const PlaylistContainer = ({ playlist, refreshPlaylist }) => {
   var yearSongs = getYearSongs(filteredTracktable)
 
 
-  const removePlaylist = async (playlistObjectId) => {
+  const removePlaylist = async (playlistObjectId, playlistId) => {
     setDeleteLoader(!deleteLoader)
     try {
-      const res = await axios.delete(`${environment}/playlists`,
-      {
-       headers: {
-         authorization: `Bearer ${cookies.access_token}`
-       },
-       data: {
-         "playlistObjectId": playlistObjectId
-       }
-     })
+      if(playlistObjectId) {
+        const res = await axios.delete(`${environment}/playlists`,
+        {
+         headers: {
+           authorization: `Bearer ${cookies.access_token}`
+         },
+         data: {
+           "playlistObjectId": playlistObjectId
+         }
+       })
+      }
 
-     deletePlaylistInfoFromLocalStorage(playlistObjectId)
+    // delete by playlistId since playlist isn't neccesarily saved by user in db
+     deletePlaylistInfoFromLocalStorage(playlistId)
      setDeleteLoader(false)
      navigate('/')
     } catch (err) {
@@ -98,7 +102,10 @@ const PlaylistContainer = ({ playlist, refreshPlaylist }) => {
         />
       </div>
       <div className='flex-container'>
-        <About removePlaylist={() => removePlaylist(playlistObjectId)} deleteLoader={deleteLoader}/>
+        <About
+          removePlaylist={() => removePlaylist(playlistObjectId, playlistId)}
+          deleteLoader={deleteLoader}
+        />
       </div>
     </>
   )
